[FEAT] 챌린지 조회 API 부분 구현

+++ src/main/java/com/BeilsangServer/domain/challenge/dto/ChallengeRequestDTO.java
@@ -23,10 +23,10 @@ public class ChallengeRequestDTO {
         private String title; // 챌린지 제목
         private LocalDate startDate; // 챌린지 시작 날짜
         private ChallengePeriod period; // 챌린지 기간(일주일/한달)
-        private Long totalGoalDay; // 목표 실천 일수
+        private Integer totalGoalDay; // 목표 실천 일수
         private Category category; // 카테고리
         private String details; // 챌린지 세부설명
         private List<String> notes; // 챌린지 유의사항
-        private Long joinPoint; // 참여 포인트
+        private Integer joinPoint; // 참여 포인트
     }
 }
